libs/cloud_storage/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "secondbrain-cloud-storage"
version = "0.1.0"
description = "Cloud storage abstraction layer for SecondBrain"
dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth>=2.0.0",
    "httpx>=0.25.0",
]
